success flag

+++ backend/src/main/java/com/mcmasterbaja/analyzer/Analyzer.java
@@ -1,15 +1,5 @@
 package com.mcmasterbaja.analyzer;
 
-import com.mcmasterbaja.annotations.OnAnalyzerException;
-import com.mcmasterbaja.exceptions.InvalidColumnException;
-import com.mcmasterbaja.exceptions.InvalidInputFileException;
-import com.mcmasterbaja.exceptions.InvalidOutputFileException;
-import com.mcmasterbaja.model.AnalyzerParams;
-import com.opencsv.CSVReader;
-import com.opencsv.CSVReaderBuilder;
-import com.opencsv.CSVWriter;
-import com.opencsv.CSVWriterBuilder;
-import com.opencsv.ICSVWriter;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.FileReader;
@@ -22,6 +12,17 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
+
+import com.mcmasterbaja.annotations.OnAnalyzerException;
+import com.mcmasterbaja.exceptions.InvalidColumnException;
+import com.mcmasterbaja.exceptions.InvalidInputFileException;
+import com.mcmasterbaja.exceptions.InvalidOutputFileException;
+import com.mcmasterbaja.model.AnalyzerParams;
+import com.opencsv.CSVReader;
+import com.opencsv.CSVReaderBuilder;
+import com.opencsv.CSVWriter;
+import com.opencsv.CSVWriterBuilder;
+import com.opencsv.ICSVWriter;
 
 @OnAnalyzerException
 public abstract class Analyzer {
@@ -49,8 +50,7 @@ public abstract class Analyzer {
   }
 
   public void getReaders(String[] filePaths, Consumer<Map<String, CSVReader>> action) {
-    Map<String, CSVReader> readersMap =
-        new LinkedHashMap<>(); // LinkedHashMap to preserve insertion order
+    Map<String, CSVReader> readersMap = new LinkedHashMap<>(); // LinkedHashMap to preserve insertion order
 
     try {
       for (String filePath : filePaths) {
@@ -86,8 +86,7 @@ public abstract class Analyzer {
       // If any exceptions occurred while closing readers, combine them and throw as a
       // single exception
       if (!exceptions.isEmpty()) {
-        InvalidInputFileException combo =
-            new InvalidInputFileException("Multiple exceptions occurred.");
+        InvalidInputFileException combo = new InvalidInputFileException("Multiple exceptions occurred.");
         for (Exception e : exceptions) {
           combo.addSuppressed(e);
         }
@@ -98,7 +97,8 @@ public abstract class Analyzer {
   }
 
   /**
-   * Default behaviour is to use the 0th output file, will need to be overridden in some special
+   * Default behaviour is to use the 0th output file, will need to be overridden
+   * in some special
    * cases
    *
    * @return Filename of the analyzer output
@@ -108,23 +108,27 @@ public abstract class Analyzer {
   }
 
   public void getWriter(String filePath, Consumer<ICSVWriter> action) {
+    boolean success = false;
     try (FileWriter fileWriter = new FileWriter(filePath);
         BufferedWriter bufferedWriter = new BufferedWriter(fileWriter);
-        ICSVWriter writer =
-            new CSVWriterBuilder(bufferedWriter)
-                .withSeparator(CSVWriter.DEFAULT_SEPARATOR)
-                .withQuoteChar(CSVWriter.NO_QUOTE_CHARACTER)
-                .withEscapeChar(CSVWriter.DEFAULT_ESCAPE_CHARACTER)
-                .withLineEnd(CSVWriter.DEFAULT_LINE_END)
-                .build()) {
+        ICSVWriter writer = new CSVWriterBuilder(bufferedWriter)
+            .withSeparator(CSVWriter.DEFAULT_SEPARATOR)
+            .withQuoteChar(CSVWriter.NO_QUOTE_CHARACTER)
+            .withEscapeChar(CSVWriter.DEFAULT_ESCAPE_CHARACTER)
+            .withLineEnd(CSVWriter.DEFAULT_LINE_END)
+            .build()) {
       action.accept(writer);
+      success = true;
     } catch (IOException e) {
-      try {
-        Files.deleteIfExists(Paths.get(filePath));
-      } catch (IOException delException) {
-        e.addSuppressed(delException);
-      }
       throw new InvalidOutputFileException("Failed to write to output file: " + filePath, e);
+    } finally {
+      if (!success) {
+        try {
+          Files.deleteIfExists(Paths.get(filePath));
+        } catch (IOException deleteException) {
+          throw new InvalidOutputFileException("Failed to delete output file: " + filePath, deleteException);
+        }
+      }
     }
   }
